safety net

+++ resources/lib/epgprocess.py
@@ -20,26 +20,34 @@ class voyo_epg(threading.Thread):
         self.__workdir = workdir
         if workdir[len(workdir)-1] != '/':
             self.__workdir += '/'
+        self.__xmlepg = '{0}epg.xml'.format(self.__workdir)
+        self.__gzfile = '{0}epg.xml.gz'.format(self.__workdir)
 
     def __download(self, chunk_size=128):
-        self.__gzfile = '{0}epg.xml.gz'.format(self.__workdir)
+        requests.packages.urllib3.disable_warnings()
         attempt = 0
         while attempt < 5:
             attempt += 1
-            r = requests.get(self.__url, stream=True)
-            if r.status_code == 200:
-                with open(self.__gzfile, 'wb') as fd:
-                    for chunk in r.iter_content(chunk_size=chunk_size):
-                        fd.write(chunk)
-                return True
+            try:
+                r = requests.get(self.__url, stream=True, allow_redirects=True, verify=False)
+                if r.status_code == 200:
+                    with open(self.__gzfile, 'wb') as fd:
+                        for chunk in r.iter_content(chunk_size=chunk_size):
+                            fd.write(chunk)
+                    return True
+            except:
+                return False
         return False
 
     def __unpack(self):
-        self.__xmlepg = '{0}epg.xml'.format(self.__workdir)
         with gzip.open(self.__gzfile, "rb") as f:
-            content = f.read()
-            with open(self.__xmlepg, 'wb') as xmlf:
-                xmlf.write(content)
+            try:
+                content = f.read()
+                with open(self.__xmlepg, 'wb') as xmlf:
+                    xmlf.write(content)
+                    return True
+            except:
+                return False
 
     def __getLogo(self):
         logos = {}
@@ -86,8 +94,10 @@ class voyo_epg(threading.Thread):
         return epg_dict
 
     def __tidyup(self):
-        os.unlink(self.__xmlepg)
-        os.unlink(self.__gzfile)
+        if os.path.exists(self.__xmlepg):
+            os.unlink(self.__xmlepg)
+        if os.path.exists(self.__gzfile):
+            os.unlink(self.__gzfile)
 
     def run(self):
         epg_exists = False
@@ -98,29 +108,29 @@ class voyo_epg(threading.Thread):
             now = time.time()
             epg_exists = True
         if not epg_exists or mtime + 24*60*60 < now: # more that 24 hours - expired
-            if self.__download():
-                self.__unpack()
+            if self.__download() and self.__unpack():
                 logostr = json.dumps(self.__getLogo(), ensure_ascii=False)
                 with open(logofname, 'w') as f:
                     f.write(logostr)
                 epg_str = json.dumps(self.__getChannelEpg(), ensure_ascii=False)
                 with codecs.open(epgfname, 'w', 'utf-8') as f:
                     f.write(epg_str)
-                self.__tidyup()
+            self.__tidyup()
 
 def main():
     epg = voyo_epg('./')
     epg.start()
     epg.join()
-    with codecs.open('epg.json', 'r', 'utf-8') as f:
-        epgs = f.read()
-        epgd = json.loads(epgs)
-        for e in epgd:
-            print('\n==> {0}\n'.format(e))
-            eepglst = epgd[e]
-            for it in eepglst:
-                print('{0} - {1} {2}'.format(
-                    it[0], it[1], it[2].encode('utf-8')))
+    if os.path.exists('epg.json'):
+        with codecs.open('epg.json', 'r', 'utf-8') as f:
+            epgs = f.read()
+            epgd = json.loads(epgs)
+            for e in epgd:
+                print('\n==> {0}\n'.format(e))
+                eepglst = epgd[e]
+                for it in eepglst:
+                    print('{0} - {1} {2}'.format(
+                        it[0], it[1], it[2].encode('utf-8')))
 
 
 if __name__ == '__main__':
